Aili最后一次提交

+++ FineGoods/ButtonTableViewCell.m
@@ -236,7 +236,7 @@
 }
 
 - (void)buttonAction:(UIButton *)button{
-    NSLog(@"%ld",button.tag);
+    NSLog(@"%ld",(long)button.tag);
 
 }
